+++ Back/app/api/tipologia.js
@@ -17,9 +17,11 @@ module.exports = function(app){
     api.salvaTipologia = (req, res) => {
         const knex = app.conexao.conexaoBDKnex();
         const tipologia = req.body;
+        const logDAO = new app.infra.LogDAO(knex);
         
         knex('tipologia').insert(tipologia)
             .then(resultado => {
+                logDAO.logTipologia(req.headers['cod_usuario'], 'tipologia', 'i', null, resultado[0]);
                 knex.destroy();
                 res.status(200).end();
             })
@@ -35,19 +37,32 @@ module.exports = function(app){
         const { cod_tipologia } = req.params;
 
         if(cod_tipologia){
-            const knex = app.conexao.conexaoBDKnex();
             const tipologia = req.body;
-    
-            knex('tipologia').where('cod_tipologia', cod_tipologia).update(tipologia)
-                .then(resultado => {
-                    knex.destroy();
-                    res.status(200).end();
-                })
-                .catch(erro => {
+            const connection = app.conexao.conexaoBD();
+            const tipologiaDAO = new app.infra.TipologiaDAO(connection);
+            tipologiaDAO.listarTipologiaLog (cod_tipologia, (erro, resultado) =>{
+                if(erro){
                     console.log(erro);
-                    knex.destroy();
                     res.status(500).send(app.api.erroPadrao());
-                });
+                }
+                else{
+                    const knex = app.conexao.conexaoBDKnex();
+                    const logDAO = new app.infra.LogDAO (knex);
+                    let espelho = resultado[0].espelho;
+                    knex('tipologia').where('cod_tipologia', cod_tipologia).update(tipologia)
+                        .then(resultado => {
+                            logDAO.logTipologia(req.headers['cod_usuario'], 'tipologia', 'u', espelho, cod_tipologia);
+                            knex.destroy();
+                            res.status(200).end();
+                        })
+                        .catch(erro => {
+                            console.log(erro);
+                            knex.destroy();
+                            res.status(500).send(app.api.erroPadrao());
+                        });
+                }
+            });
+            connection.end();
         } else { res.status(400).send(app.api.erroPadrao()); }
     }
     
@@ -56,22 +71,35 @@ module.exports = function(app){
         const { cod_tipologia } = req.params;
 
         if(cod_tipologia){
-            const knex = app.conexao.conexaoBDKnex();
-
-            knex('tipologia').where('cod_tipologia', cod_tipologia).delete()
-                .then(resultado => {
-                    knex.destroy();
-                    res.status(200).end();
-                })
-                .catch(erro => {
+            const connection = app.conexao.conexaoBD();
+            const tipologiaDAO = new app.infra.TipologiaDAO(connection);
+            tipologiaDAO.listarTipologiaLog(cod_tipologia, (erro, resultado) => {
+                if(erro){
                     console.log(erro);
-                    knex.destroy();
-                    if(erro.errno == 1451){
-                        res.status(500).send('Esta tipologia não pode ser apagada pois existem outras informações associadas a ela.');
-                    } else {
-                        res.status(500).send(app.api.erroPadrao());
-                    }
-                });
+                    res.status(500).send(app.api.erroPadrao());
+                }
+                else{
+                    const knex = app.conexao.conexaoBDKnex();
+                    const logDAO = new app.infra.LogDAO(knex);
+                    let espelho = resultado[0].espelho;
+                    knex('tipologia').where('cod_tipologia', cod_tipologia).delete()
+                        .then(resultado => {
+                            logDAO.logTipologia(req.headers['cod_usuario'], 'tipologia', 'd', espelho, cod_tipologia);
+                            knex.destroy();
+                            res.status(200).end();
+                        })
+                        .catch(erro => {
+                            console.log(erro);
+                            knex.destroy();
+                            if(erro.errno == 1451){
+                                res.status(500).send('Esta tipologia não pode ser apagada pois existem outras informações associadas a ela.');
+                            } else {
+                                res.status(500).send(app.api.erroPadrao());
+                            }
+                        });
+                }
+            });
+            connection.end();
         } else { res.status(400).send(app.api.erroPadrao()); }
     }
